+++ R/utils_stat_test_label.R
@@ -176,3 +176,31 @@ keep_only_tbl_df_classes <- function(x){
   }
   x
 }
+
+
+
+# The dot-dot notation (`..p.signif..`) was deprecated in ggplot2 3.4.0.
+# after_stat(p.signif) should be used. This function makes automatic
+# conversion if user specified ..p.signif..
+convert_label_dotdot_notation_to_after_stat <- function(mapping){
+  if (!is.null(mapping) ){
+    label <- mapping$label
+    if (!is.null(mapping$label)) {
+
+      label <-  rlang::as_label(mapping$label)
+      dot_dot_labels <- c(
+        "p.signif", "p.adj.signif", "p.format", "p", "p.adj",
+        "eq.label", "adj.rr.label", "rr.label", "AIC.label", "BIC.label"
+      )
+      for (dot_dot_label in dot_dot_labels) {
+        label <- gsub(
+          pattern = paste0("..", dot_dot_label, ".."),
+          replacement = paste0("ggplot2::after_stat(", dot_dot_label, ")"),
+          x = label, fixed = TRUE
+        )
+      }
+      mapping$label <- parse(text = label)[[1]]
+    }
+  }
+  mapping
+}